+++ mockrxandroidble/src/main/java/com/polidea/rxandroidble2/mockrxandroidble/RxBleClientMock.java
@@ -6,6 +6,7 @@ import android.bluetooth.BluetoothGattDescriptor;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
+import com.polidea.rxandroidble2.scan.IsConnectable;
 import com.polidea.rxandroidble2.RxBleClient;
 import com.polidea.rxandroidble2.RxBleDevice;
 import com.polidea.rxandroidble2.RxBleScanResult;
@@ -39,6 +40,7 @@ public class RxBleClientMock extends RxBleClient {
 
         private ReplaySubject<RxBleDeviceMock> discoverableDevicesSubject;
         private Set<RxBleDevice> bondedDevices;
+        private Set<RxBleDevice> connectedPeripherals;
 
         /**
          * Build a new {@link RxBleClientMock}.
@@ -46,6 +48,7 @@ public class RxBleClientMock extends RxBleClient {
         public Builder() {
             this.discoverableDevicesSubject = ReplaySubject.create();
             this.bondedDevices = new HashSet<>();
+            this.connectedPeripherals = new HashSet<>();
         }
 
         public Builder setDeviceDiscoveryObservable(@NonNull Observable<RxBleDeviceMock> discoverableDevicesObservable) {
@@ -70,6 +73,16 @@ public class RxBleClientMock extends RxBleClient {
          */
         public Builder addBondedDevice(@NonNull RxBleDevice rxBleDevice) {
             bondedDevices.add(rxBleDevice);
+            return this;
+        }
+
+        /**
+         * Add a {@link RxBleDevice} to the list of connected devices.
+         *
+         * @param rxBleDevice device that the mocked client should contain. Use {@link RxBleDeviceMock.Builder} to create them.
+         */
+        public Builder addConnectedPeripheral(@NonNull RxBleDevice rxBleDevice) {
+            connectedPeripherals.add(rxBleDevice);
             return this;
         }
 
@@ -248,10 +261,12 @@ public class RxBleClientMock extends RxBleClient {
     }
 
     private Set<RxBleDevice> bondedDevices;
+    private Set<RxBleDevice> connectedPeripherals;
     private ReplaySubject<RxBleDeviceMock> discoveredDevicesSubject;
 
     private RxBleClientMock(Builder builder) {
         bondedDevices = builder.bondedDevices;
+        connectedPeripherals = builder.connectedPeripherals;
         discoveredDevicesSubject = builder.discoverableDevicesSubject;
     }
 
@@ -271,6 +286,11 @@ public class RxBleClientMock extends RxBleClient {
     @Override
     public Set<RxBleDevice> getBondedDevices() {
         return bondedDevices;
+    }
+
+    @Override
+    public Set<RxBleDevice> getConnectedPeripherals() {
+        return connectedPeripherals;
     }
 
     @Override
@@ -362,7 +382,9 @@ public class RxBleClientMock extends RxBleClient {
                 rssi,
                 System.currentTimeMillis() * 1000000,
                 ScanCallbackType.CALLBACK_TYPE_FIRST_MATCH,
-                scanRecord);
+                scanRecord,
+                IsConnectable.LEGACY_UNKNOWN
+        );
     }
 
     private static boolean maskedDataEquals(@NonNull byte[] data1, @NonNull byte[] data2, @Nullable byte[] mask) {
@@ -409,7 +431,17 @@ public class RxBleClientMock extends RxBleClient {
     }
 
     @Override
+    public boolean isConnectRuntimePermissionGranted() {
+        return true;
+    }
+
+    @Override
     public String[] getRecommendedScanRuntimePermissions() {
+        return new String[0];
+    }
+
+    @Override
+    public String[] getRecommendedConnectRuntimePermissions() {
         return new String[0];
     }
 }