+++ rxandroidble/src/main/java/com/polidea/rxandroidble2/internal/RxBleDeviceImpl.java
@@ -12,13 +12,12 @@ import com.polidea.rxandroidble2.exceptions.BleAlreadyConnectedException;
 import com.polidea.rxandroidble2.internal.connection.Connector;
 
 import com.polidea.rxandroidble2.internal.logger.LoggerUtil;
-import java.util.concurrent.Callable;
+import com.polidea.rxandroidble2.internal.util.CheckerConnectPermission;
+
 import java.util.concurrent.atomic.AtomicBoolean;
 
 import bleshadow.javax.inject.Inject;
 import io.reactivex.Observable;
-import io.reactivex.ObservableSource;
-import io.reactivex.functions.Action;
 
 @DeviceScope
 class RxBleDeviceImpl implements RxBleDevice {
@@ -26,17 +25,20 @@ class RxBleDeviceImpl implements RxBleDevice {
     final BluetoothDevice bluetoothDevice;
     final Connector connector;
     private final BehaviorRelay<RxBleConnection.RxBleConnectionState> connectionStateRelay;
+    private final CheckerConnectPermission checkerConnectPermission;
     final AtomicBoolean isConnected = new AtomicBoolean(false);
 
     @Inject
     RxBleDeviceImpl(
             BluetoothDevice bluetoothDevice,
             Connector connector,
-            BehaviorRelay<RxBleConnection.RxBleConnectionState> connectionStateRelay
+            BehaviorRelay<RxBleConnection.RxBleConnectionState> connectionStateRelay,
+            CheckerConnectPermission checkerConnectPermission
     ) {
         this.bluetoothDevice = bluetoothDevice;
         this.connector = connector;
         this.connectionStateRelay = connectionStateRelay;
+        this.checkerConnectPermission = checkerConnectPermission;
     }
 
     @Override
@@ -69,20 +71,12 @@ class RxBleDeviceImpl implements RxBleDevice {
     }
 
     public Observable<RxBleConnection> establishConnection(final ConnectionSetup options) {
-        return Observable.defer(new Callable<ObservableSource<RxBleConnection>>() {
-            @Override
-            public ObservableSource<RxBleConnection> call() {
-                if (isConnected.compareAndSet(false, true)) {
-                    return connector.prepareConnection(options)
-                            .doFinally(new Action() {
-                                @Override
-                                public void run() {
-                                    isConnected.set(false);
-                                }
-                            });
-                } else {
-                    return Observable.error(new BleAlreadyConnectedException(bluetoothDevice.getAddress()));
-                }
+        return Observable.defer(() -> {
+            if (isConnected.compareAndSet(false, true)) {
+                return connector.prepareConnection(options)
+                        .doFinally(() -> isConnected.set(false));
+            } else {
+                return Observable.error(new BleAlreadyConnectedException(bluetoothDevice.getAddress()));
             }
         });
     }
@@ -90,6 +84,13 @@ class RxBleDeviceImpl implements RxBleDevice {
     @Override
     @Nullable
     public String getName() {
+        return getName(false);
+    }
+
+    private String getName(boolean placeholderIfNoPermission) {
+        if (placeholderIfNoPermission && !checkerConnectPermission.isConnectRuntimePermissionGranted()) {
+            return "[NO BLUETOOTH_CONNECT PERMISSION]";
+        }
         return bluetoothDevice.getName();
     }
 
@@ -125,7 +126,7 @@ class RxBleDeviceImpl implements RxBleDevice {
     public String toString() {
         return "RxBleDeviceImpl{"
                 + LoggerUtil.commonMacMessage(bluetoothDevice.getAddress())
-                + ", name=" + bluetoothDevice.getName()
+                + ", name=" + getName(true)
                 + '}';
     }
 }