+++ rxandroidble/src/test/groovy/com/polidea/rxandroidble2/RxBleClientTest.groovy
@@ -2,39 +2,38 @@ package com.polidea.rxandroidble2
 
 import android.bluetooth.BluetoothDevice
 import android.content.Context
-import android.os.Build
 import com.polidea.rxandroidble2.exceptions.BleScanException
 
 import com.polidea.rxandroidble2.internal.RxBleDeviceProvider
 import com.polidea.rxandroidble2.internal.operations.Operation
 import com.polidea.rxandroidble2.internal.scan.*
 import com.polidea.rxandroidble2.internal.serialization.ClientOperationQueue
+import com.polidea.rxandroidble2.internal.util.CheckerConnectPermission
 import com.polidea.rxandroidble2.internal.util.CheckerScanPermission
 import com.polidea.rxandroidble2.internal.util.ClientStateObservable
 import com.polidea.rxandroidble2.internal.util.ScanRecordParser
 import com.polidea.rxandroidble2.scan.BackgroundScanner
 import com.polidea.rxandroidble2.scan.ScanSettings
-import hkhc.electricspock.ElectricSpecification
+import spock.lang.Specification
 import io.reactivex.Observable
 import io.reactivex.ObservableSource
 import io.reactivex.ObservableTransformer
 import io.reactivex.annotations.NonNull
 import io.reactivex.observers.TestObserver
 import io.reactivex.schedulers.TestScheduler
-import org.robolectric.annotation.Config
 import spock.lang.Unroll
 
 import static com.polidea.rxandroidble2.exceptions.BleScanException.*
 
 @SuppressWarnings("GrDeprecatedAPIUsage")
-@Config(manifest = Config.NONE, constants = BuildConfig, sdk = Build.VERSION_CODES.LOLLIPOP)
-class RxBleClientTest extends ElectricSpecification {
+class RxBleClientTest extends Specification {
 
     BackgroundScanner backgroundScanner = Mock(BackgroundScanner)
     DummyOperationQueue dummyQueue = new DummyOperationQueue()
     RxBleClient objectUnderTest
     Context contextMock = Mock Context
     ScanRecordParser scanRecordParserSpy = Spy ScanRecordParser
+    MockBluetoothManagerWrapper bluetoothManagerWrapperSpy = Spy MockBluetoothManagerWrapper
     MockRxBleAdapterWrapper bleAdapterWrapperSpy = Spy MockRxBleAdapterWrapper
     MockRxBleAdapterStateObservable adapterStateObservable = Spy MockRxBleAdapterStateObservable
     MockLocationServicesStatus locationServicesStatusMock = Spy MockLocationServicesStatus
@@ -53,7 +52,8 @@ class RxBleClientTest extends ElectricSpecification {
     ScanSetup mockScanSetup = new ScanSetup(mockOperationScan, mockObservableTransformer)
     ScanPreconditionsVerifier mockScanPreconditionVerifier = Mock ScanPreconditionsVerifier
     InternalToExternalScanResultConverter mockMapper = Mock InternalToExternalScanResultConverter
-    CheckerScanPermission mockCheckerLocationPermission = Mock CheckerScanPermission
+    CheckerScanPermission mockCheckerScanPermission = Mock CheckerScanPermission
+    CheckerConnectPermission mockCheckerConnectPermission = Mock CheckerConnectPermission
     private static someUUID = UUID.randomUUID()
     private static otherUUID = UUID.randomUUID()
     private static Date suggestedDateToRetry = new Date()
@@ -77,6 +77,7 @@ class RxBleClientTest extends ElectricSpecification {
         mockOperationScan.run(_) >> Observable.never()
         mockScanSetupBuilder.build(_, _) >> mockScanSetup
         objectUnderTest = new RxBleClientImpl(
+                bluetoothManagerWrapperSpy,
                 bleAdapterWrapperSpy,
                 queue,
                 adapterStateObservable.asObservable(),
@@ -90,7 +91,8 @@ class RxBleClientTest extends ElectricSpecification {
                 new TestScheduler(),
                 Mock(ClientComponent.ClientComponentFinalizer),
                 backgroundScanner,
-                mockCheckerLocationPermission
+                mockCheckerScanPermission,
+                mockCheckerConnectPermission
         )
     }
 
@@ -103,6 +105,18 @@ class RxBleClientTest extends ElectricSpecification {
 
         when:
         def results = objectUnderTest.getBondedDevices()
+
+        then:
+        assert results.size() == 2
+    }
+
+    def "should return connected devices"() {
+        given:
+        bluetoothPeripheralConnected("AA:AA:AA:AA:AA:AA")
+        bluetoothPeripheralConnected("BB:BB:BB:BB:BB:BB")
+
+        when:
+        def results = objectUnderTest.getConnectedPeripherals()
 
         then:
         assert results.size() == 2
@@ -224,7 +238,7 @@ class RxBleClientTest extends ElectricSpecification {
         1 * bleAdapterWrapperSpy.stopLegacyLeScan(_)
 
         and:
-        scanSubscription.assertTerminated()
+        scanSubscription.assertError(BleScanException.class)
     }
 
     def "should stop scan after all subscribers are unsubscribed"() {
@@ -483,6 +497,13 @@ class RxBleClientTest extends ElectricSpecification {
         bleAdapterWrapperSpy.addBondedDevice(mock)
     }
 
+    def bluetoothPeripheralConnected(String address) {
+        def mock = Mock(BluetoothDevice)
+        mock.getAddress() >> address
+        mock.hashCode() >> address.hashCode()
+        bluetoothManagerWrapperSpy.addConnectedPeripheral(mock)
+    }
+
     def "should throw UnsupportedOperationException if .getBleDevice() is called on system that has no Bluetooth capabilities"() {
 
         given:
@@ -522,13 +543,13 @@ class RxBleClientTest extends ElectricSpecification {
     }
 
     @Unroll
-    def "should pass call to CheckerLocationPermission when called .isScanRuntimePermissionGranted() and proxy back the result"() {
+    def "should pass call to CheckerScanPermission when called .isScanRuntimePermissionGranted() and proxy back the result"() {
 
         when:
         def result = objectUnderTest.isScanRuntimePermissionGranted()
 
         then:
-        1 * mockCheckerLocationPermission.isScanRuntimePermissionGranted() >> expectedResult
+        1 * mockCheckerScanPermission.isScanRuntimePermissionGranted() >> expectedResult
 
         and:
         result == expectedResult
@@ -537,7 +558,23 @@ class RxBleClientTest extends ElectricSpecification {
         expectedResult << [true, false]
     }
 
-    def "should pass call to CheckerLocationPermission when called .getRecommendedScanRuntimePermissions() and proxy back the result"() {
+    @Unroll
+    def "should pass call to CheckerConnectPermission when called .isConnectRuntimePermissionGranted() and proxy back the result"() {
+
+        when:
+        def result = objectUnderTest.isConnectRuntimePermissionGranted()
+
+        then:
+        1 * mockCheckerConnectPermission.isConnectRuntimePermissionGranted() >> expectedResult
+
+        and:
+        result == expectedResult
+
+        where:
+        expectedResult << [true, false]
+    }
+
+    def "should pass call to CheckerScanPermission when called .getRecommendedScanRuntimePermissions() and proxy back the result"() {
 
         given:
         String[] resultRef = new String[0]
@@ -546,7 +583,22 @@ class RxBleClientTest extends ElectricSpecification {
         def result = objectUnderTest.getRecommendedScanRuntimePermissions()
 
         then:
-        1 * mockCheckerLocationPermission.getRecommendedScanRuntimePermissions() >> resultRef
+        1 * mockCheckerScanPermission.getRecommendedScanRuntimePermissions() >> resultRef
+
+        and:
+        result == resultRef
+    }
+    
+    def "should pass call to CheckerConnectPermission when called .getRecommendedConnectRuntimePermissions() and proxy back the result"() {
+       
+        given:
+        String[] resultRef = new String[0]
+
+        when:
+        def result = objectUnderTest.getRecommendedConnectRuntimePermissions()
+
+        then:
+        1 * mockCheckerConnectPermission.getRecommendedConnectRuntimePermissions() >> resultRef
 
         and:
         result == resultRef